requirements.txt
fastapi
uvicorn
typer[all]
requests
appium-python-client
requests
python-dotenv

